+++ src/main/kotlin/com/group4/ticketingservice/repository/ReservationRepository.kt
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface ReservationRepository : JpaRepository<Reservation, Long>
+interface ReservationRepository : JpaRepository<Reservation, Int>
